requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
whitenoise==4.1.2